// src/types/nav2017.ts
export interface NAVCustomer {
  No: string;
  Name: string;
  Name2?: string;
  Address: string;
  Address2?: string;
  City: string;
  County?: string;
  PostCode: string;
  CountryRegionCode: string;
  PhoneNo?: string;
  Email?: string;
  HomePage?: string;
  ContactName?: string;
  SalespersonCode?: string;
  CustomerPriceGroup?: string;
  CustomerDiscountGroup?: string;
  PaymentTermsCode?: string;
  PaymentMethodCode?: string;
  CurrencyCode?: string;
  CreditLimitLCY?: number;
  BalanceLCY?: number;
  BalanceDueLCY?: number;
  SalesLCY?: number;
  ProfitLCY?: number;
  InvDiscountsLCY?: number;
  PmtDiscountsLCY?: number;
  PmtToleranceLCY?: number;
  OutstandingOrdersLCY?: number;
  ShippedNotInvoicedLCY?: number;
  ApplicationMethod?: number;
  PricesIncludingVAT?: boolean;
  GenBusPostingGroup?: string;
  VATBusPostingGroup?: string;
  CustomerPostingGroup?: string;
  Blocked?: number;
  LastDateModified?: string;
  DateFilter?: string;
  GlobalDimension1Filter?: string;
  GlobalDimension2Filter?: string;
  LocationFilter?: string;
  CurrencyFilter?: string;
}